+++ lib/app/core/models/novel_meta.dart
@@ -1,9 +1,11 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter/rendering.dart';
 import 'package:hb_db/hb_db.dart';
 import 'package:than_pkg/than_pkg.dart';
+import 'package:uuid/uuid.dart';
 
 class NovelMetaAdapter extends HBAdapter<NovelMeta> {
   @override
@@ -24,6 +26,7 @@ class NovelMetaAdapter extends HBAdapter<NovelMeta> {
 
 class NovelMeta {
   static final String metaName = 'meta.json';
+  final String id;
   final String title;
   final String author;
   final String translator;
@@ -37,7 +40,9 @@ class NovelMeta {
   final int readed;
   final String originalTitle;
   final String englishTitle;
+  final DateTime date;
   NovelMeta({
+    required this.id,
     required this.title,
     required this.author,
     required this.translator,
@@ -51,15 +56,17 @@ class NovelMeta {
     required this.readed,
     required this.originalTitle,
     required this.englishTitle,
+    required this.date,
   });
 
   factory NovelMeta.createEmpty() {
     return NovelMeta.create();
   }
 
-  factory NovelMeta.create() {
+  factory NovelMeta.create({String title = 'Untitled', String? id}) {
     return NovelMeta(
-      title: 'Untitled',
+      id: id ?? Uuid().v4(),
+      title: title,
       originalTitle: '',
       englishTitle: '',
       author: 'Unknown',
@@ -72,16 +79,21 @@ class NovelMeta {
       pageUrls: [],
       tags: [],
       readed: 0,
+      date: DateTime.now(),
     );
   }
 
   static Future<NovelMeta> fromPath(String novelPath) async {
     final metaFile = File(pathJoin(novelPath, metaName));
+    final novelDir = Directory(novelPath);
 
     if (metaFile.existsSync()) {
       final source = await metaFile.readAsString();
-      if (source.isEmpty) return NovelMeta.create();
-      return NovelMeta.fromMap(jsonDecode(source));
+      if (source.isEmpty) return NovelMeta.create(title: novelPath.getName());
+      return NovelMeta.fromMap(
+        jsonDecode(source),
+        novelDirMiliDateNumber: novelDir.getDate.millisecondsSinceEpoch,
+      );
     } else {
       // read old files
       return await _getOldConfig(novelPath);
@@ -105,71 +117,32 @@ class NovelMeta {
     return [];
   }
 
-  NovelMeta copyWith({
-    String? title,
-    String? originalTitle,
-    String? englishTitle,
-    String? author,
-    String? translator,
-    String? mc,
-    String? desc,
-    List<String>? otherTitleList,
-    bool? isAdult,
-    bool? isCompleted,
-    List<String>? pageUrls,
-    List<String>? tags,
-    int? readed,
+  factory NovelMeta.fromMap(
+    Map<String, dynamic> map, {
+    int novelDirMiliDateNumber = 0,
   }) {
+    final id = map.getString(['id'], def: Uuid().v4());
+    final date = map.getInt(['date'], def: novelDirMiliDateNumber);
+    final otherTitleList = map['otherTitleList'] ?? [];
+    final pageUrls = map['pageUrls'] ?? [];
+    final tags = map['tags'] ?? [];
     return NovelMeta(
-      title: title ?? this.title,
-      originalTitle: originalTitle ?? this.originalTitle,
-      englishTitle: englishTitle ?? this.englishTitle,
-      author: author ?? this.author,
-      translator: translator ?? this.translator,
-      mc: mc ?? this.mc,
-      desc: desc ?? this.desc,
-      otherTitleList: otherTitleList ?? this.otherTitleList,
-      isAdult: isAdult ?? this.isAdult,
-      isCompleted: isCompleted ?? this.isCompleted,
-      pageUrls: pageUrls ?? this.pageUrls,
-      tags: tags ?? this.tags,
-      readed: readed ?? this.readed,
-    );
-  }
-
-  Map<String, dynamic> toMap() {
-    return <String, dynamic>{
-      'title': title,
-      'originalTitle': originalTitle,
-      'englishTitle': englishTitle,
-      'author': author,
-      'translator': translator,
-      'mc': mc,
-      'desc': desc,
-      'otherTitleList': otherTitleList,
-      'isAdult': isAdult,
-      'isCompleted': isCompleted,
-      'pageUrls': pageUrls,
-      'tags': tags,
-      'readed': readed,
-    };
-  }
-
-  factory NovelMeta.fromMap(Map<String, dynamic> map) {
-    return NovelMeta(
-      title: map.getString(['title']),
-      originalTitle: map.getString(['originalTitle']),
-      englishTitle: map.getString(['englishTitle']),
-      author: map.getString(['author']),
-      translator: map.getString(['translator']),
-      mc: map.getString(['mc']),
-      desc: map.getString(['desc']),
-      otherTitleList: List<String>.from((map['otherTitleList'] ?? [])),
+      id: id,
+      title: map['title'] as String,
+      author: map['author'] as String,
+      translator: map['translator'] as String,
+      mc: map['mc'] as String,
+      desc: map['desc'] as String,
+      otherTitleList: List<String>.from(otherTitleList),
       isAdult: map['isAdult'] as bool,
       isCompleted: map['isCompleted'] as bool,
-      pageUrls: List<String>.from((map['pageUrls'] ?? [])),
-      tags: List<String>.from((map['tags'] ?? [])),
+      pageUrls: List<String>.from(pageUrls),
+      tags: List<String>.from(tags),
       readed: map['readed'] as int,
+      originalTitle: map.getString(['originalTitle']),
+      englishTitle: map.getString(['englishTitle']),
+      date: DateTime.fromMillisecondsSinceEpoch(date),
+      //get from novel folder -> date
     );
   }
 
@@ -185,12 +158,13 @@ class NovelMeta {
     final tagsFile = File(pathJoin(novelPath, 'tags'));
 
     int readed = 0;
-    final title = novelPath.getName();
+    final dirName = novelPath.getName();
     if (readedFile.existsSync()) {
       readed = int.tryParse(await readedFile.readAsString()) ?? 0;
     }
     final meta = NovelMeta(
-      title: title,
+      id: Uuid().v4(),
+      title: dirName,
       originalTitle: '',
       englishTitle: '',
       author: authorFile.existsSync()
@@ -211,6 +185,7 @@ class NovelMeta {
           ? _getListFromString(await tagsFile.readAsString())
           : [],
       readed: readed,
+      date: DateTime.now(),
     );
     // delete
     if (mcFile.existsSync()) {
@@ -243,5 +218,61 @@ class NovelMeta {
     await meta.save(novelPath);
 
     return meta;
+  }
+
+  NovelMeta copyWith({
+    String? id,
+    String? title,
+    String? author,
+    String? translator,
+    String? mc,
+    String? desc,
+    List<String>? otherTitleList,
+    bool? isAdult,
+    bool? isCompleted,
+    List<String>? pageUrls,
+    List<String>? tags,
+    int? readed,
+    String? originalTitle,
+    String? englishTitle,
+    DateTime? date,
+  }) {
+    return NovelMeta(
+      id: id ?? this.id,
+      title: title ?? this.title,
+      author: author ?? this.author,
+      translator: translator ?? this.translator,
+      mc: mc ?? this.mc,
+      desc: desc ?? this.desc,
+      otherTitleList: otherTitleList ?? this.otherTitleList,
+      isAdult: isAdult ?? this.isAdult,
+      isCompleted: isCompleted ?? this.isCompleted,
+      pageUrls: pageUrls ?? this.pageUrls,
+      tags: tags ?? this.tags,
+      readed: readed ?? this.readed,
+      originalTitle: originalTitle ?? this.originalTitle,
+      englishTitle: englishTitle ?? this.englishTitle,
+      date: date ?? this.date,
+    );
+  }
+
+  Map<String, dynamic> toMap() {
+    return <String, dynamic>{
+      'id': id,
+      'title': title,
+      'author': author,
+      'translator': translator,
+      'mc': mc,
+      'desc': desc,
+      'otherTitleList': otherTitleList,
+      'isAdult': isAdult,
+      'isCompleted': isCompleted,
+      'pageUrls': pageUrls,
+      'tags': tags,
+      'readed': readed,
+      'originalTitle': originalTitle,
+      'englishTitle': englishTitle,
+      'date': date.millisecondsSinceEpoch,
+    };
   }
 }