+++ lib/app/core/services/novel_services.dart
@@ -2,8 +2,10 @@ import 'dart:io';
 
 import 'package:novel_v3/app/core/extensions/novel_extension.dart';
 import 'package:novel_v3/app/core/models/novel.dart';
+import 'package:novel_v3/app/core/models/novel_meta.dart';
 import 'package:novel_v3/more_libs/setting/core/path_util.dart';
 import 'package:than_pkg/extensions/file_system_entity_extension.dart';
+import 'package:uuid/uuid.dart';
 
 class NovelServices {
   static Future<List<Novel>> getAll() async {
@@ -22,10 +24,15 @@ class NovelServices {
     return list;
   }
 
-  static Future<Novel?> createNovelWithTitle(String title) async {
-    final dir = Directory(PathUtil.getSourcePath(name: title));
-    if (dir.existsSync()) return null;
-    await dir.create();
-    return await Novel.fromPath(dir.path);
+  static Future<Novel> createNovelFolder({required NovelMeta meta}) async {
+    final name = Uuid().v4();
+    final dir = Directory(PathUtil.getSourcePath(name: name));
+    if (!dir.existsSync()) {
+      await dir.create();
+    }
+    final newMeta = meta.copyWith(id: name);
+    await newMeta.save(dir.path);
+
+    return Novel.create(name: name, path: dir.path, meta: newMeta);
   }
 }