+++ lib/app/others/novl_db/novl_export_list_tile.dart
@@ -38,7 +38,7 @@ class _NovlExportListTileState extends State<NovlExportListTile> {
               progressManager: NovlExportProgressManager(
                 novel: widget.novel,
                 info: info,
-                outPath: PathUtil.getOutPath(name: widget.novel.title),
+                outPath: PathUtil.getOutPath(name: widget.novel.meta.title),
                 onDone: widget.onClosed,
               ),
             ),