+++ lib/app/others/novl_db/novl_install_progress_manager.dart
@@ -1,9 +1,9 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'dart:async';
 import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:hb_db/hb_db.dart';
+import 'package:novel_v3/app/others/novl_db/novl_data.dart';
 import 'package:t_widgets/progress_manager/progress_manager_interface.dart';
 import 'package:t_widgets/progress_manager/progress_message.dart';
 import 'package:than_pkg/utils/index.dart';
@@ -11,8 +11,10 @@ import 'package:than_pkg/utils/index.dart';
 class NovlInstallProgressManager extends ProgressManagerInterface {
   final List<DBFEntry> installList;
   final String installNovelPath;
+  final NovlData novl;
   final VoidCallback? onDone;
   NovlInstallProgressManager({
+    required this.novl,
     required this.installList,
     required this.installNovelPath,
     this.onDone,
@@ -47,6 +49,10 @@ class NovlInstallProgressManager extends ProgressManagerInterface {
           },
         );
       }
+
+      // override novel meta
+      final meta = novl.novelMeta.copyWith(date: DateTime.now());
+      await meta.save(installNovelPath);
 
       streamController.add(ProgressMessage.done());
       await streamController.close();